+++ CHANGELOG.md
@@ -1,6 +1,10 @@
 # Attess Changelog
-> Current Version: 0.0.1-alpha
+> Current Version: 0.1.1-alpha
+
+# 0.1.1-alpha
+- Feature: Accounts module
+- Feature: Module banner
 
 # 0.0.1-alpha / Jan 15th 2023
 - Project setup
-- Account module released
+- Feature: Account module released